Implemented Sending / Receiving TEBucks :)

See merge request campuses/pgh/jan-2022/java-blue/student-pairs/java-blue-capstone-module-2-team-3!4

+++ tenmo-client/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -4,23 +4,23 @@ import java.math.BigDecimal;
 
 public class Account {
 
-    private int account_id;
-    private int user_id; // might want to be a User object
+    private long account_id;
+    private long user_id; // might want to be a User object
     private BigDecimal balance;
 
-    public int getAccount_id() {
+    public long getAccount_id() {
         return account_id;
     }
 
-    public void setAccount_id(int account_id) {
+    public void setAccount_id(long account_id) {
         this.account_id = account_id;
     }
 
-    public int getUser_id() {
+    public long getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(int user_id) {
+    public void setUser_id(long user_id) {
         this.user_id = user_id;
     }
